+++ crates/block-producer/src/block_builder/mod.rs
@@ -1,4 +1,4 @@
-use std::sync::Arc;
+use std::{collections::BTreeSet, sync::Arc};
 
 use async_trait::async_trait;
 use miden_node_utils::formatting::{format_array, format_blake3_digest};
@@ -76,18 +76,41 @@ where
         let updated_accounts: Vec<_> =
             batches.iter().flat_map(TransactionBatch::updated_accounts).collect();
 
-        let created_notes = batches.iter().map(TransactionBatch::created_notes).cloned().collect();
+        let created_notes: Vec<_> =
+            batches.iter().map(TransactionBatch::output_notes).cloned().collect();
 
         let produced_nullifiers: Vec<Nullifier> =
             batches.iter().flat_map(TransactionBatch::produced_nullifiers).collect();
+
+        let created_notes_set: BTreeSet<_> = created_notes
+            .iter()
+            .flat_map(|batch| batch.iter().map(|note| note.id()))
+            .collect();
+
+        let dangling_notes: BTreeSet<_> = batches
+            .iter()
+            .flat_map(TransactionBatch::unauthenticated_input_notes)
+            .filter(|&note_id| !created_notes_set.contains(note_id))
+            .copied()
+            .collect();
 
         let block_inputs = self
             .store
             .get_block_inputs(
                 updated_accounts.iter().map(BlockAccountUpdate::account_id),
                 produced_nullifiers.iter(),
+                dangling_notes.iter(),
             )
             .await?;
+
+        if block_inputs.found_unauthenticated_notes.len() < dangling_notes.len() {
+            return Err(BuildBlockError::UnauthenticatedNotesNotFound(
+                dangling_notes
+                    .difference(&block_inputs.found_unauthenticated_notes)
+                    .copied()
+                    .collect(),
+            ));
+        }
 
         let block_header_witness = BlockWitness::new(block_inputs, batches)?;