+++ crates/store/src/db/migrations/001-init.sql
@@ -30,11 +30,12 @@ CREATE TABLE
     note_type   INTEGER NOT NULL, -- 1-Public (0b01), 2-OffChain (0b10), 3-Encrypted (0b11)
     sender      INTEGER NOT NULL,
     tag         INTEGER NOT NULL,
+    aux         INTEGER NOT NULL,
     merkle_path BLOB    NOT NULL,
     details     BLOB,
 
     PRIMARY KEY (block_num, batch_index, note_index),
-    CONSTRAINT fk_block_num FOREIGN KEY (block_num) REFERENCES block_headers (block_num),
+    CONSTRAINT fk_block_num FOREIGN KEY (block_num) REFERENCES block_headers(block_num),
     CONSTRAINT notes_type_in_enum CHECK (note_type BETWEEN 1 AND 3),
     CONSTRAINT notes_block_num_is_u32 CHECK (block_num BETWEEN 0 AND 0xFFFFFFFF),
     CONSTRAINT notes_batch_index_is_u32 CHECK (batch_index BETWEEN 0 AND 0xFFFFFFFF),
@@ -50,7 +51,7 @@ CREATE TABLE
     details      BLOB,
 
     PRIMARY KEY (account_id),
-    CONSTRAINT fk_block_num FOREIGN KEY (block_num) REFERENCES block_headers (block_num),
+    CONSTRAINT fk_block_num FOREIGN KEY (block_num) REFERENCES block_headers(block_num),
     CONSTRAINT accounts_block_num_is_u32 CHECK (block_num BETWEEN 0 AND 0xFFFFFFFF)
 ) STRICT, WITHOUT ROWID;
 
@@ -62,8 +63,23 @@ CREATE TABLE
     block_num        INTEGER NOT NULL,
 
     PRIMARY KEY (nullifier),
-    CONSTRAINT fk_block_num FOREIGN KEY (block_num) REFERENCES block_headers (block_num),
+    CONSTRAINT fk_block_num FOREIGN KEY (block_num) REFERENCES block_headers(block_num),
     CONSTRAINT nullifiers_nullifier_is_digest CHECK (length(nullifier) = 32),
     CONSTRAINT nullifiers_nullifier_prefix_is_u16 CHECK (nullifier_prefix BETWEEN 0 AND 0xFFFF),
     CONSTRAINT nullifiers_block_num_is_u32 CHECK (block_num BETWEEN 0 AND 0xFFFFFFFF)
 ) STRICT, WITHOUT ROWID;
+
+CREATE TABLE
+    transactions
+(
+    transaction_id BLOB    NOT NULL,
+    account_id     INTEGER NOT NULL,
+    block_num      INTEGER NOT NULL,
+
+    PRIMARY KEY (transaction_id),
+    CONSTRAINT fk_block_num FOREIGN KEY (block_num) REFERENCES block_headers(block_num),
+    CONSTRAINT transactions_block_num_is_u32 CHECK (block_num BETWEEN 0 AND 0xFFFFFFFF)
+) STRICT, WITHOUT ROWID;
+
+CREATE INDEX idx_transactions_account_id ON transactions(account_id);
+CREATE INDEX idx_transactions_block_num ON transactions(block_num);