+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+* Add `init` command for node and faucet (#392).
+* Added crate to distribute node RPC protobuf files (#391).
+* Changed state sync endpoint to return a list of `TransactionSummary` objects instead of just transaction IDs (#386).
+* Fixed faucet note script so that it uses the `aux` input (#387).
+* Added `aux` column to notes table (#384).
+* Standardised CI and Makefile across Miden repositories (#367)
+* Removed client dependency from faucet (#368).
+* Changed sync endpoint to return a list of committed transactions (#377).
+* Added support for unauthenticated transaction notes (#390).
+
 ## 0.3.0 (2024-05-15)
 
 * Added option to mint pulic notes in the faucet (#339).